+++ fantoch/src/protocol/mod.rs
@@ -50,7 +50,7 @@ pub trait Protocol: Clone {
     fn new(
         process_id: ProcessId,
         config: Config,
-    ) -> (Self, Vec<(Self::PeriodicEvent, usize)>);
+    ) -> (Self, Vec<(Self::PeriodicEvent, u64)>);
 
     fn id(&self) -> ProcessId;
